pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flake8>=7.2.0",
    "mcp[cli]>=1.6.0",
    "openai>=1.75.0",
    "python-dotenv>=0.21.0",
]
